This is a major commit adding the following functionalities to the code: * a fully parallel implementation of the linear algebra, inverters and update; * a new input file reader; * the geometry parameters can now be specified from the input file.
New *_field types have been introduced together with the existing spinor_field. All of these types now have an internal mechanism to keep track of the MPI non-blocking communications used in the code. Some global variables have new names and others have been added for the use in the parallel code (like PID, i.e. the current Process ID, or the MPI communicators). 

While the RHMC update seems to work, almost all the test in the Directory TestProgram are broken at the moment. Also the codes in PureGauge, QuenchedSpectrum and Observables (in part) are not yet functional. Also the update seems to be not correct when an odd local lattice size is used (maybe a problem of the geometry or its initialization). The precision issue with the inverters still needs a solution. The implementation of the boundary conditions seems incorrect. A lot more work is needed to make the code reliable and user friendly!

The new input file reader should be easier to use and allows for more flexibility than the fixed input format reader used in the non-mpi branch at the moment.

Many other smaller improvements, e.g. a faster algorithm in main.C for the remez code, an enable/disable feature in the logger.

A copyright notice has been added to the source files.

Enjoy!


git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-mpi@86 37a69121-b030-0410-a54a-fdca83ffee47