The new geometry: the geometry has now a totally different layout that allows to allocate an even or odd spinor at time. The layout of the geometry is organized in this way:
|--bulk even--|--3d border even--|--3d buffer even--|--bulk odd--|--3d border odd--|--3d buffer odd--|--2d border even--|--2d border odd--|--2d buffer even--|--2dbuffer odd--|
|------------------------------------------------------------------glattice---------------------------------------------------------------------------------------------------|
|-----------------------glat_even-------------------|---------------glat_odd-------------------------|

In order to accomodate the same indexing among glattice glat_even and glat_odd the new variable shift has been introduced, it keeps in account of the different point of start of glat_odd.
All the tests of the geometry have been implemented and checked.
All the other test codes have been verified, and they fully work.
Fixed many issues with the intel compiler about float/double conversion.
The RHMC on a comparison run gives the correct plaquette.

BEWARE THE HMC SEEMS NOT TO WORK!!!!
BEWARE THE HMC SEEMS NOT TO WORK!!!!
BEWARE THE HMC SEEMS NOT TO WORK!!!!

Not checked also the wilson loop the spectrum and the modenumber routines.
This release is not meant to be used for production until fully verified.
Antonio






git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-HMC@477 37a69121-b030-0410-a54a-fdca83ffee47