Added replica mode to code.
It is now posible to run multiple copies of the code inside one MPI run.
The number of replicas can be specified for each direction in the input file, e.g. REP_T = 2, etc.

When replicas are used each replica runs inside a directory named "Rep_#" where #=0, 1, ... is the Replica ID.
Different input files should be given inside the replica directories, and a global input file, including the geometry and replica informations must be given in the parent directory.




git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-HMC@482 37a69121-b030-0410-a54a-fdca83ffee47