Added FGMRES and GMRES inverters.
Added DDalphaAMG.c file in inverters that contain the functions necessary for the DDalphaAMG inverter. The functions are not done yet. However the V-cycle will now just call the SAP, so tests can be performed for the structure of the code.

Added a test program for the DDalphaAMG inverter. So far it seems to work wil SAP alone in place of the "V-cycle". 




git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-MG@532 37a69121-b030-0410-a54a-fdca83ffee47