I have now made some of the key linear algebra functions for the small spinors, which are to be used in the small spinor inverter. I have also made a Minimal Residual Inverter, but haven't tested it yet. In the end I think we want to use GMRES.
In my previous commit I used only small spinors defined as static in the DDalphaAMG file. This was in order to save some of all the allocation/freeing calls. Now I have found that this was not very nice to work with so from now on I want to pass the pointers around. At this point the code is somewhere in between the two ideas, but this will be resolved very soon.




git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/Branches/HiRep-MG@537 37a69121-b030-0410-a54a-fdca83ffee47