Modified HMC update and input file structure. See Doc/action for a detailed explanation.
Summary of changes:
* A new monomial structure has been added to the code
* it is now possible to change the action from the input file, by specifying any number of monomials   
* the MD integrator structure has also been changed: it is now possible to specify the multilevel integrator from input file and to assign each monomial the desired integrator level

An example of the new format for the input file is in the HMC directory.

WARNING: please note that now in the input file the parameter 'run name' is the WHOLE name of the configuration file except for the suffix 'nXXX'!

!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
NB: this version has not extensively tested and should NOT be used for production yet.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!



git-svn-id: file:///Users/pica/Downloads/GitClone/svnrep/suN/HiRep@677 37a69121-b030-0410-a54a-fdca83ffee47